auto/requirements.txt
gunicorn==20.0.4
Pillow==5.1.0
pipenv==2018.11.26
django-crispy-forms==1.8.0
django_phonenumber_field==4.0.0
Django==2.2.6
dj_database_url==0.5.0
djangorestframework==3.11.0
python-decouple==3.3
phonenumbers==8.11.1
whitenoise==5.0.1
psycopg2-binary==2.8.5


